working qt interaface...
